fixed some bugs

+++ lib/resources/shop/bloc/models/articles/article_by_id_model.dart
@@ -56,7 +56,7 @@ class Product with _$Product {
     required String? pickupAddress,
     required String? description,
     required int? regularPrice,
-    required int? salePrice,
+    required double? salePrice,
     required int? discount,
     required String? color,
     required List<String>? images,
